Changed script to load plugins via vundle
Previously all the plugins were handled manually with git pulls. This file uses vundle to load the plugins.
To get it to work you need to install vundle into ~/.config/nvim/pack/bundle.
So when that exist you can run :PluginUpdate and vundle should update and download the other plugins.